Make AppGrabber one window
