+++ code/Cairo Desktop/CairoDesktop.AppGrabber/AppGrabberUI.xaml
@@ -118,7 +118,150 @@
                     Grid.Row="6"
                     Grid.Column="3">
                     <Button Style="{StaticResource CairoButtonStyle}"
-                        Click="Button_Click">Continue</Button>
+                        Click="btnContinue_Click">Continue</Button>
+                </StackPanel>
+            </Grid>
+        </Border>
+        <Border Margin="12,7,12,17" Name="bdrPage2" Visibility="Collapsed">
+            <Grid>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="32" />
+                    <RowDefinition Height="40" />
+                    <RowDefinition Height="*" />
+                    <RowDefinition Height="32" />
+                    <RowDefinition Height="30" />
+                    <RowDefinition Height="32" />
+                </Grid.RowDefinitions>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="32" />
+                    <ColumnDefinition Width="330" />
+                    <ColumnDefinition Width="24" />
+                    <ColumnDefinition Width="330" />
+                    <ColumnDefinition Width="32" />
+                </Grid.ColumnDefinitions>
+                <StackPanel Grid.Row="1"
+                Grid.Column="1"
+                Grid.RowSpan="2">
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF000000"
+                    FontSize="22"
+                    FontWeight="Bold"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Categories
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Categories allow you to group applications by type so they are easier to find. Once you finish, you can add, remove, or edit categories and applications through the Programs menu itself. For convenience, you may also run the App Grabber from the Cairo menu.
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF000000"
+                    FontSize="18"
+                    Margin="0,20,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Directions:
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Drag and drop applications into the categories you desire.
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Click the plus button to add a new category.
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Click the X button on a category to delete it. (Applications in a deleted category will become uncategorized.)
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Right-click a category to rename it.
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Click the "Hide" button to show only applications that need categorization.
+                    </TextBlock>
+                    <TextBlock FontFamily="Corbel"
+                    Foreground="#FF777777"
+                    FontSize="14"
+                    Margin="0,10,0,0"
+                    Width="330"
+                    TextWrapping="Wrap">
+                                    Items placed in the Quick Launch category will appear in the quick launch area of the Cairo taskbar. Items placed here will not be removed from the Programs menu.  To remove an item, drag it out from the Quick Launch category.
+                    </TextBlock>
+                </StackPanel>
+                <Border BorderThickness="1"
+                BorderBrush="#FF888888"
+                Grid.Row="1"
+                Grid.Column="3">
+                    <DockPanel Width="330" DockPanel.Dock="Top" Style="{StaticResource AppGrabberListBoxHeaderStyle}">
+                        <Button DockPanel.Dock="Left"
+                            Click="btnAdd_Click"
+                            FontSize="18"
+                            FontFamily="Tahoma"
+                            Style="{StaticResource AppGrabberListBoxHeaderButtonStyle}"
+                            Margin="5,0,0,0">
+                            +
+                        </Button>
+                        <ToggleButton DockPanel.Dock="Right" Name="tglHide" Click="tglHide_Click" Style="{StaticResource AppGrabberListBoxHeaderTextToggleButtonStyle}" Margin="0,0,6,0" IsChecked="{Binding bAppsHidden, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type Window}}, UpdateSourceTrigger=PropertyChanged}">
+                            Hide
+                        </ToggleButton>
+                        <TextBlock Style="{StaticResource AppGrabberListBoxHeaderTextStyle}"
+                            DockPanel.Dock="Top">
+                                        Program Categories
+                        </TextBlock>
+                    </DockPanel>
+                </Border>
+                <Border BorderThickness="1,0,1,1"
+                BorderBrush="#FF888888"
+                Grid.Row="2"
+                Grid.Column="3">
+                    <ScrollViewer x:Name="scrollViewer" UseLayoutRounding="True"
+                    HorizontalScrollBarVisibility="Hidden"
+                    VerticalScrollBarVisibility="Auto"
+                    CanContentScroll="True"
+                    Background="White"
+                    DragOver="ScrollViewer_DragOver">
+                        <ItemsControl x:Name="categoryView"
+                        Style="{StaticResource CategoryDisplayStyle}" VirtualizingStackPanel.VirtualizationMode="Recycling" ScrollViewer.HorizontalScrollBarVisibility="Hidden" Visibility="Hidden" />
+                    </ScrollViewer>
+                </Border>
+                <StackPanel HorizontalAlignment="Right"
+                Grid.Row="4"
+                Grid.Column="1">
+                    <Button Style="{StaticResource CairoButtonStyle}"
+                    Click="SkipWizard">Cancel</Button>
+                </StackPanel>
+                <StackPanel HorizontalAlignment="Left"
+                Grid.Row="4"
+                Grid.Column="3">
+                    <Button Style="{StaticResource CairoButtonStyle}"
+                    Click="btnSave_Click">Finish</Button>
                 </StackPanel>
             </Grid>
         </Border>
