+++ code/Cairo Desktop/CairoDesktop.AppGrabber/AppGrabberUI.xaml.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Forms;
 
 namespace CairoDesktop.AppGrabber
@@ -13,6 +14,33 @@ namespace CairoDesktop.AppGrabber
     {
         private AppGrabber appGrabber;
         ObservableCollection<ApplicationInfo> programsMenuAppsCollection;
+
+        public Visibility AppVisibility
+        {
+            get { return (Visibility)GetValue(AppVisibilityProperty); }
+            set { SetValue(AppVisibilityProperty, value); }
+        }
+
+        private bool bAppsHidden
+        {
+            get { return (AppVisibility != Visibility.Visible); }
+            set
+            {
+                if (value)
+                {
+                    SetValue(AppVisibilityProperty, Visibility.Collapsed);
+                    SetValue(bAppsHiddenProperty, true);
+                }
+                else
+                {
+                    SetValue(AppVisibilityProperty, Visibility.Visible);
+                    SetValue(bAppsHiddenProperty, false);
+                }
+            }
+        }
+
+        public static readonly DependencyProperty AppVisibilityProperty = DependencyProperty.Register("AppVisibility", typeof(Visibility), typeof(AppGrabberUI), new PropertyMetadata(null));
+        private static readonly DependencyProperty bAppsHiddenProperty = DependencyProperty.Register("bAppsHidden", typeof(bool), typeof(AppGrabberUI), new PropertyMetadata(null));
 
         public AppGrabberUI()
             : this(new AppGrabber())
@@ -27,58 +55,22 @@ namespace CairoDesktop.AppGrabber
             Height = SystemParameters.MaximizedPrimaryScreenHeight - 100;
         }
 
-        void programsMenuAppsCollection_CollectionChanged(object sender, System.Collections.Specialized.NotifyCollectionChangedEventArgs e) {
-            if (e.Action == System.Collections.Specialized.NotifyCollectionChangedAction.Remove) {
-                foreach (ApplicationInfo app in e.OldItems) {
-                    if (app.Category != null) {
-                        app.Category.Remove(app);
-                    }
-                }
-            }
-        }
+        #region Button Clicks
 
         private void SkipWizard(object sender, RoutedEventArgs e)
         {
             this.Close();
         }
 
-        private void Button_Click(object sender, RoutedEventArgs e) {
-            this.Visibility = Visibility.Hidden;
-            AppGrabber.ui2Instance = new AppGrabberUI_Page2(this.appGrabber, programsMenuAppsCollection);
-            AppGrabber.ui2Instance.Show();
-            this.Close();
+        private void btnContinue_Click(object sender, RoutedEventArgs e)
+        {
+            goPage2();
         }
 
-        private void Window_ContentRendered(object sender, EventArgs e)
+        private void btnSave_Click(object sender, RoutedEventArgs e)
         {
-            // Now to add them to the list on the AppGrabber
-            // Create ObservableCollections to bind to the ListViews
-
-            ObservableCollection<ApplicationInfo> installedAppsCollection = new ObservableCollection<ApplicationInfo>();
-            programsMenuAppsCollection = appGrabber.CategoryList.FlatList;
-            InstalledAppsView.ItemsSource = installedAppsCollection;
-            ProgramsMenuAppsView.ItemsSource = programsMenuAppsCollection;
-            // Need to use an event handler to remove Apps from categories when moved to the "Installed Applications" listing
-            programsMenuAppsCollection.CollectionChanged += new System.Collections.Specialized.NotifyCollectionChangedEventHandler(programsMenuAppsCollection_CollectionChanged);
-
-            // Grab the Programs
-            List<ApplicationInfo> apps = appGrabber.ProgramList;
-
-            // Iterate thru the apps, creating ApplicationInfoPanels and
-            // add them to the installedAppsCollection
-            foreach (ApplicationInfo app in apps)
-            {
-                if (!programsMenuAppsCollection.Contains(app))
-                {
-                    installedAppsCollection.Add(app);
-                }
-            }
-            AppViewSorter.Sort(installedAppsCollection, "Name");
-            AppViewSorter.Sort(programsMenuAppsCollection, "Name");
-
-            // show content
-            bdrLoad.Visibility = Visibility.Collapsed;
-            bdrMain.Visibility = Visibility.Visible;
+            appGrabber.Save();
+            this.Close();
         }
 
         private void btnBrowse_Click(object sender, RoutedEventArgs e)
@@ -114,6 +106,38 @@ namespace CairoDesktop.AppGrabber
             }
         }
 
+        private void btnAdd_Click(object sender, RoutedEventArgs e)
+        {
+            appGrabber.CategoryList.Add(new Category("New"));
+            scrollViewer.ScrollToEnd();
+        }
+
+        private void tglHide_Click(object sender, RoutedEventArgs e)
+        {
+            if (!bAppsHidden)
+                bAppsHidden = true;
+            else
+                bAppsHidden = false;
+        }
+
+        #endregion
+
+        #region Events
+
+        void programsMenuAppsCollection_CollectionChanged(object sender, System.Collections.Specialized.NotifyCollectionChangedEventArgs e)
+        {
+            if (e.Action == System.Collections.Specialized.NotifyCollectionChangedAction.Remove)
+            {
+                foreach (ApplicationInfo app in e.OldItems)
+                {
+                    if (app.Category != null)
+                    {
+                        app.Category.Remove(app);
+                    }
+                }
+            }
+        }
+
         private void ProgramsMenuAppsView_MouseDoubleClick(object sender, System.Windows.Input.MouseButtonEventArgs e)
         {
             ApplicationInfo app = ((FrameworkElement)e.OriginalSource).DataContext as ApplicationInfo;
@@ -132,9 +156,98 @@ namespace CairoDesktop.AppGrabber
             (ProgramsMenuAppsView.ItemsSource as ObservableCollection<ApplicationInfo>).Add(app);
         }
 
+        private void Window_ContentRendered(object sender, EventArgs e)
+        {
+            // Now to add them to the list on the AppGrabber
+            // Create ObservableCollections to bind to the ListViews
+
+            ObservableCollection<ApplicationInfo> installedAppsCollection = new ObservableCollection<ApplicationInfo>();
+            programsMenuAppsCollection = appGrabber.CategoryList.FlatList;
+            InstalledAppsView.ItemsSource = installedAppsCollection;
+            ProgramsMenuAppsView.ItemsSource = programsMenuAppsCollection;
+            // Need to use an event handler to remove Apps from categories when moved to the "Installed Applications" listing
+            programsMenuAppsCollection.CollectionChanged += new System.Collections.Specialized.NotifyCollectionChangedEventHandler(programsMenuAppsCollection_CollectionChanged);
+
+            // Grab the Programs
+            List<ApplicationInfo> apps = appGrabber.ProgramList;
+
+            // Iterate thru the apps, creating ApplicationInfoPanels and
+            // add them to the installedAppsCollection
+            foreach (ApplicationInfo app in apps)
+            {
+                if (!programsMenuAppsCollection.Contains(app))
+                {
+                    installedAppsCollection.Add(app);
+                }
+            }
+            AppViewSorter.Sort(installedAppsCollection, "Name");
+            AppViewSorter.Sort(programsMenuAppsCollection, "Name");
+
+            // show content
+            bdrLoad.Visibility = Visibility.Collapsed;
+            bdrMain.Visibility = Visibility.Visible;
+        }
+
         private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
         {
             AppGrabber.uiInstance = null;
+        }
+
+        private void ScrollViewer_DragOver(object sender, System.Windows.DragEventArgs e)
+        {
+            int scrollTollerance = 12;
+            ScrollViewer scroller = sender as ScrollViewer;
+            System.Diagnostics.Debug.WriteLine("Mouse rel scroller:" + e.GetPosition(scroller).Y);
+            System.Diagnostics.Debug.WriteLine("Scroller Height:" + scroller.ActualHeight);
+            if (e.GetPosition(scroller).Y > scroller.ActualHeight - scrollTollerance)
+            {
+                scroller.LineDown();
+            }
+            else if (e.GetPosition(scroller).Y < scrollTollerance)
+            {
+                scroller.LineUp();
+            }
+        }
+
+        #endregion
+
+        private void goPage2()
+        {
+            bdrMain.Visibility = Visibility.Collapsed;
+            bdrPage2.Visibility = Visibility.Visible;
+
+            // Grab the Programs
+            CategoryList catList = appGrabber.CategoryList;
+
+            // Get the Uncategorized category - create it if it doesn't exist.
+            Category uncat = catList.GetCategory("Uncategorized");
+            if (uncat == null)
+            {
+                uncat = new Category("Uncategorized");
+                uncat.ShowInMenu = false;
+                catList.Add(uncat);
+            }
+
+            // Get the Quick Launch category - create it if it doesn't exist.
+            Category quicklaunch = catList.GetCategory("Quick Launch");
+            if (quicklaunch == null)
+            {
+                quicklaunch = new Category("Quick Launch");
+                quicklaunch.ShowInMenu = false;
+                catList.Add(quicklaunch);
+            }
+
+            // Add Apps to uncat if they haven't been added to a category yet.
+            foreach (ApplicationInfo app in programsMenuAppsCollection)
+            {
+                if (app.Category == null)
+                {
+                    uncat.Add(app);
+                }
+            }
+
+            categoryView.ItemsSource = catList;
+            categoryView.Visibility = Visibility.Visible;
         }
     }
 }
